+++ src/internal.hpp
@@ -570,7 +570,7 @@ struct Internal {
   // Forward reasoning through propagation in 'propagate.cpp'.
   //
   int assignment_level (int lit, Clause *);
-  void build_chain_for_units (int lit, Clause *reason);
+  void build_chain_for_units (int lit, Clause *reason, bool forced);
   void build_chain_for_empty ();
   void search_assign (int lit, Clause *);
   void search_assign_driving (int lit, Clause *reason);
@@ -597,6 +597,7 @@ struct Internal {
   //
   void learn_empty_clause ();
   void learn_unit_clause (int lit);
+  void learn_external_propagated_unit_clause (int lit);
   void bump_variable (int lit);
   void bump_variables ();
   int recompute_glue (Clause *);