+++ src/lratchecker.cpp
@@ -111,9 +111,9 @@ void LratChecker::collect_garbage_clauses () {
 /*------------------------------------------------------------------------*/
 
 LratChecker::LratChecker (Internal *i)
-    : internal (i), size_vars (0), num_clauses (0), num_garbage (0),
-      size_clauses (0), clauses (0), garbage (0), last_hash (0),
-      last_id (0) {
+    : internal (i), size_vars (0), num_clauses (0), num_finalized (0),
+      num_garbage (0), size_clauses (0), clauses (0), garbage (0),
+      last_hash (0), last_id (0) {
   LOG ("LRAT CHECKER new");
 
   // Initialize random number table for hash function.
@@ -356,20 +356,6 @@ bool LratChecker::check (vector<uint64_t> proof_chain) {
     }
     if (!unit) {
       LOG ("LRAT CHECKER check succeded, clause falsified %" PRIu64, id);
-      // TODO fuzz with this to see where errors are.
-      // known problems with lratbuilder? not interesting anymore...
-      // fails in decompose... I do not have a trace for instantiate but I
-      // am sure it also fails this assert
-      // assert (proof_chain.back () == id || internal->opts.lratexternal ||
-      //        internal->opts.decompose);
-      // || internal->opts.instantiate);
-      // assert (proof_chain.back () == id);    // also tempting since this
-      // basically means the proof chain
-      // is unnecessarily long.
-      // but unfortunatly this also
-      // fails when we prove the
-      // inconsistent clause to justify
-      // whatever
       checking = true;
       break;
     }
@@ -390,8 +376,7 @@ bool LratChecker::check (vector<uint64_t> proof_chain) {
 
 void LratChecker::add_original_clause (uint64_t id, const vector<int> &c) {
   START (checking);
-  LOG (c, "LRAT CHECKER addition of original clause");
-  LOG ("LRAT CHECKER clause id %" PRIu64, id);
+  LOG (c, "LRAT CHECKER addition of original clause[%" PRIu64 "]", id);
   stats.added++;
   stats.original++;
   import_clause (c);
@@ -405,8 +390,7 @@ void LratChecker::add_original_clause (uint64_t id, const vector<int> &c) {
 void LratChecker::add_derived_clause (uint64_t id, const vector<int> &c,
                                       const vector<uint64_t> &proof_chain) {
   START (checking);
-  LOG (c, "LRAT CHECKER addition of derived clause");
-  LOG ("LRAT CHECKER clause id %" PRIu64, id);
+  LOG (c, "LRAT CHECKER addition of derived clause[%" PRIu64 "]", id);
   stats.added++;
   stats.derived++;
   import_clause (c);
@@ -427,8 +411,7 @@ void LratChecker::add_derived_clause (uint64_t id, const vector<int> &c,
 
 void LratChecker::add_derived_clause (uint64_t id, const vector<int> &c) {
   START (checking);
-  LOG (c, "LRAT CHECKER addition of derived unproven clause");
-  LOG ("LRAT CHECKER clause id %" PRIu64, id);
+  LOG (c, "LRAT CHECKER checking derived unproven clause[%" PRIu64 "]", id);
   stats.added++;
   import_clause (c);
   last_id = id;
@@ -450,8 +433,7 @@ void LratChecker::add_derived_clause (uint64_t id, const vector<int> &c) {
 
 void LratChecker::delete_clause (uint64_t id, const vector<int> &c) {
   START (checking);
-  LOG (c, "LRAT CHECKER checking deletion of clause");
-  LOG ("LRAT CHECKER clause id %" PRIu64, id);
+  LOG (c, "LRAT CHECKER checking deletion of clause[%" PRIu64 "]", id);
   stats.deleted++;
   import_clause (c);
   last_id = id;
@@ -500,6 +482,62 @@ void LratChecker::delete_clause (uint64_t id, const vector<int> &c) {
   STOP (checking);
 }
 
+
+void LratChecker::finalize_clause (uint64_t id, const vector<int> &c) {
+  START (checking);
+  LOG (c, "LRAT CHECKER checking finalize of clause[%" PRIu64 "]", id);
+  stats.finalized++;
+  num_finalized++;
+  import_clause (c);
+  last_id = id;
+  LratCheckerClause **p = find (id), *d = *p;
+  if (d) {
+    for (const auto &lit : imported_clause)
+      mark (lit) = true;
+    const int *dp = d->literals;
+    for (unsigned i = 0; i < d->size; i++) {
+      int lit = *(dp + i);
+      if (!mark (lit)) {        // should never happen since ids
+        fatal_message_start (); // are unique.
+        fputs ("deleted clause not in proof:\n", stderr);
+        for (const auto &lit : imported_clause)
+          fprintf (stderr, "%d ", lit);
+        fputc ('0', stderr);
+        fatal_message_end ();
+      }
+    }
+    for (const auto &lit : imported_clause)
+      mark (lit) = false;
+
+  } else {
+    fatal_message_start ();
+    fputs ("deleted clause not in proof:\n", stderr);
+    for (const auto &lit : imported_clause)
+      fprintf (stderr, "%d ", lit);
+    fputc ('0', stderr);
+    fatal_message_end ();
+  }
+  imported_clause.clear ();
+  STOP (checking);
+}
+
+// check if all clauses have been deleted
+void LratChecker::finalize_check () {
+  START (checking);
+  if (num_finalized == num_clauses) {
+    num_finalized = 0;
+    LOG ("LRAT CHECKER successful finalize check, all clauses have been deleted");
+  } else {
+    fatal_message_start ();
+    fputs ("finalize check failed ", stderr);
+    fprintf (stderr, "%" PRIu64, num_clauses);
+    fputs (" are not finalized", stderr);
+    fatal_message_end ();
+  }
+  STOP (checking);
+}
+
+  
 /*------------------------------------------------------------------------*/
 
 void LratChecker::dump () {